requirements.txt
aiohttp
discord.py>=2.4.0
python-dotenv